Rename *LRTest to *IT. They will be run by failsafe
