+++ nexus/nexus-core-plugins/nexus-indexer-lucene-plugin-parent/nexus-indexer-lucene-plugin/src/test/java/org/sonatype/nexus/index/DefaultIndexerManagerIT.java
@@ -25,7 +25,7 @@ import org.sonatype.nexus.proxy.repository.ProxyRepository;
 import org.sonatype.nexus.templates.repository.maven.Maven2ProxyRepositoryTemplate;
 
 // This is an IT just because it runs longer then 15 seconds
-public class DefaultIndexerManagerLRTest
+public class DefaultIndexerManagerIT
     extends AbstractIndexerManagerTest
 {
