+++ nexus/nexus-proxy/src/test/java/org/sonatype/nexus/proxy/attributes/perf/HashMapAttributeStoragePerformanceIT.java
@@ -14,8 +14,7 @@ package org.sonatype.nexus.proxy.attributes.perf;
 
 import org.junit.runner.RunWith;
 import org.sonatype.nexus.proxy.attributes.AttributeStorage;
-import org.sonatype.nexus.proxy.attributes.DefaultFSAttributeStorage;
-import org.sonatype.nexus.proxy.attributes.XStreamMarshaller;
+import org.sonatype.nexus.proxy.attributes.HashMapAttributeStorage;
 import org.sonatype.nexus.proxy.attributes.perf.internal.OrderedRunner;
 
 import com.carrotsearch.junitbenchmarks.annotation.AxisRange;
@@ -29,20 +28,25 @@ import com.carrotsearch.junitbenchmarks.annotation.BenchmarkMethodChart;
 @BenchmarkMethodChart
 @AxisRange( min = 0 )
 @RunWith( OrderedRunner.class )
-public class DefaultFSAttributeStoragePerformanceLRTest
+public class HashMapAttributeStoragePerformanceIT
     extends AttributeStoragePerformanceTestSupport
 {
+//    @Rule
+//    public MethodRule benchmarkRun = new BenchmarkRule();
 
-    // @Rule
-    // public MethodRule benchmarkRun = new BenchmarkRule();
+    private AttributeStorage attributeStorage =  new HashMapAttributeStorage();
+
+    public void setup()
+        throws Exception
+    {
+        super.setup();
+
+        test1PutAttribute();
+        test2PutAttributeX100();
+    }
 
     public AttributeStorage getAttributeStorage()
     {
-        DefaultFSAttributeStorage attributeStorage =
-            new DefaultFSAttributeStorage( applicationConfiguration,
-                new XStreamMarshaller() );
-        attributeStorage.initializeWorkingDirectory();
         return attributeStorage;
     }
-
 }
