+++ nexus/nexus-proxy/src/test/java/org/sonatype/nexus/proxy/storage/local/fs/perf/DefaultFSLocalRepositoryStoragePerformanceIT.java
@@ -56,7 +56,7 @@ import com.carrotsearch.junitbenchmarks.annotation.BenchmarkMethodChart;
 @BenchmarkHistoryChart( )
 @BenchmarkMethodChart( )
 @AxisRange( min = 0 )
-public class DefaultFSLocalRepositoryStoragePerformanceITLRTest
+public class DefaultFSLocalRepositoryStoragePerformanceIT
     extends AbstractNexusTestEnvironment
 {
     @Rule
